@@ -16,6 +16,8 @@ module Storey
       switches['schema-only'] = nil if @options[:structure_only]
       switches['no-privileges'] = nil
       switches['no-owner'] = nil
+      switches['host'] = @options[:host] if @options[:host].present?
+      switches['username'] = @options[:username] if @options[:username].present?
       switches[:file] = Shellwords.escape(@options[:file])
 
       if @options[:schemas]
@@ -25,7 +27,11 @@ module Storey
         end
       end
 
-      command_parts = ['pg_dump',
+      command_parts = []
+      if @options[:password].present?
+        command_parts << "PGPASSWORD=#{@options[:password]}"
+      end
+      command_parts += ['pg_dump',
                        Utils.command_line_switches_from(switches),
                        schemas_switches,
                        @options[:database]]
