@@ -8,8 +8,8 @@ module Storey
     end
 
     def dump
-      `#{command}`
-      raise 'Error dumping database' if $?.exitstatus == 1
+      stdout_str, stderr_str, status = Open3.capture3(command)
+      raise "Error dumping database: #{stderr_str}" if status.exitstatus != 0
     end
 
     private
@@ -27,10 +27,14 @@ module Storey
     end
 
     def command
-      @command ||= BuildsDumpCommand.execute(structure_only: true,
-                                             file: @file,
-                                             schemas: search_path,
-                                             database: database_name)
+      return @command if defined?(@command)
+      args = Storey.database_config.slice(:host, :username, :password).merge(
+        structure_only: true,
+        file: @file,
+        schemas: search_path,
+        database: database_name,
+      )
+      @command = BuildsDumpCommand.execute(args)
     end
 
   end
